avr: group instructions by category
This makes it easier to refactor things that only affect a specific group of
instructions. For example, if we want to change how the PC is handled, most
of the instructions that need to be changed are the in the control-flow group
so having them close together makes this easier.